+++ module/data/actor-journey.mjs
@@ -9,6 +9,7 @@ export default class MistEngineJourney extends MistEngineActorBase {
 
     schema.tags = new fields.StringField({ required: true, blank: true });
     schema.role = new fields.StringField({ required: true, blank: true });
+    schema.notes = new fields.StringField({ required: true, blank: true });
 
     schema.generalConsequences = new fields.SchemaField({
         name: new fields.StringField(),